easier resolution of types
